backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
marshmallow==3.21.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
bcrypt==4.1.3
pytest==8.2.2
pytest-flask==1.3.0